+++ services/workflows-service/src/alert/webhook-manager/webhook-http.service.ts
@@ -1,0 +1,5 @@
+import { HttpService } from '@nestjs/axios';
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export abstract class WebhookHttpService extends HttpService {}