+++ services/workflows-service/src/collection-flow/controllers/collection-flow.business.controller.ts
@@ -3,9 +3,11 @@ import { Public } from '@/common/decorators/public.decorator';
 import { UseTokenAuthGuard } from '@/common/guards/token-guard/use-token-auth.decorator';
 import { Controller, Get, Query } from '@nestjs/common';
 import { GetBusinessInformationDto } from '../dto/get-business-information-input.dto';
+import { ApiExcludeController } from '@nestjs/swagger';
 
 @Public()
 @UseTokenAuthGuard()
+@ApiExcludeController()
 @Controller('collection-flow/business')
 export class CollectionFlowBusinessController {
   constructor(protected readonly businessService: BusinessService) {}