+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/workflow-core
 
+## 0.6.0
+
+### Minor Changes
+
+- version bump
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.8.0
+
+## 0.5.52
+
+### Patch Changes
+
+- fix invoked at
+
 ## 0.5.51
 
 ### Patch Changes